requirements.txt
Django==3.2.0
psycopg2-binary
djangorestframework
drf-yasg==1.20.0
django-filter==22.1
python-dotenv
pytest-django==4.5.2